// src/sub-categories/types.ts
export type ResponseSubCategory = {
    sub_category_id: string;
    category_id: string;
    name: string;
    nick_name: string;
    created_at: string;
    updated_at: string;
    product: boolean;
};

export type ResultSubCategory = {
    subCategoryId: string;
    categoryId: string;
    name: string;
    nickName: string;
    createdAt: string;
    updatedAt: string;
    product: boolean;
};